Parameterized unit test.
It turned out that even with the Acceptance Test coverage, the Post method
could be given a degenerate implementation with hard-coded test values.
Only when the user name is parameterized is the desired data flow
implemented.
Incidentally, this is also a great illustration of why code coverage is
in itself an inadequate metric.